requirements.txt
bcrypt==3.2.0
cffi==1.14.4
click==7.1.2
cryptography==3.3.1
ecdsa==0.14.1
fastapi==0.63.0
h11==0.12.0
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.20
pydantic==1.7.3
pymongo==3.11.2
python-jose==3.2.0
python-multipart==0.0.5
rsa==4.7
six==1.15.0
starlette==0.13.6
uvicorn==0.13.3
